Refactor to "Improve code-ownership rules for ESIMD/Tools"
