+++ llvm/lib/SYCLLowerIR/ESIMD/LowerESIMD.cpp
@@ -13,7 +13,7 @@
 // values of integer template parameters they were instantiated with.
 //===----------------------------------------------------------------------===//
 
-#include "llvm/SYCLLowerIR/LowerESIMD.h"
+#include "llvm/SYCLLowerIR/ESIMD/LowerESIMD.h"
 
 #include "llvm/ADT/DenseMap.h"
 #include "llvm/ADT/DenseSet.h"
