+++ llvm/test/SYCLLowerIR/ESIMD/global.ll
