+++ llvm/test/SYCLLowerIR/ESIMD/lower_vec_arg_fp.ll
