+++ llvm/test/SYCLLowerIR/ESIMD/scalar_fptoui.ll
